refactor(experimental): a function to compile a transaction's address table lookups into a message (#1365)
* refactor(experimental): a function to compile a transaction's instructions into a message
## Summary

Over the wire, transaction instructions don't carry the addresses that they refer to, but rather the _indices_ of those addresses in a sidecar address table. This function takes in the ordered addresses and the instructions, and returns the instructions with the addresses replaced with indices.


## Test Plan

```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```

* refactor(experimental): a function to compile a transaction's static accounts into a message
## Summary

In this PR we simply map over the ordered accounts and output the address associated with each, but only for static accounts.

## Test Plan

```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```

* refactor(experimental): a function to compile a transaction's address table lookups into a message
## Summary

Once we have an ordered list of accounts, producing the address table lookups for a transaction is a matter of rolling them up into a `Vec<MessageAddressTableLookup>`. This PR implements that.

## Test Plan

```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```