refactor(experimental): A websocket transport (#1546)
* refactor(experimental): create a WebSocket impl package

# Summary

This is a private shim that builds a browser version and a node version of `WebSocket`. The browser version uses the builtin, and the node version uses `ws`

* refactor(experimental): A websocket transport

# Summary

This is as unopinionated a transport as we can have. It does not:

1. Autoreconnect
2. Perform periodic pings

You make one like this:

```ts
// Define a transport
const transport = createWebSocketTransport({url: '...'});

// Spawn a connection
const abortController = new AbortController();
const connection = await transport({signal: abortController.signal});
```

It simply connects when you ask it to, and returns:

1. a `send()` method
2. an async iterator that vends channel messages

```ts
// Consume messages
for await (const message of connection) {
  console.log('got a websocket message', message);
}
```

If you want to explicitly close the connection, `abort()` the controller associated with the `AbortSignal`

# Test Plan

```
cd packages/rpc-transport
pnpm test:unit:browser
pnpm test:unit:node
```