refactor(experimental): add read and write functions to codecs-data-structures (#1910)
Apologies for the big PR. It basically does everything we've been doing in the previous PRs of this stack but inside the `codecs-data-structures` package which contains lots of composable codecs.

On top of that, it refactors the `Set<T>` and `Map<K,V>` codecs such that they rely on the `Array<T>` codec by using `mapCodec`. Before, they were almost copy/pasted implementation which made little sense.