Validate that the public key generated from createKeyPairFromBytes() belongs to the private key (#2329)
This PR addresses #2289 

updated `createKeyPairFromBytes`. Adds a validation check to verify the public and private key by signing data and verifying that the signed data matches the public key.

Adds error code for public/private key mismatches.

Included in this PR is testcases to test the possible errors within these functions and creating a successful keypair from a 64-byte array.